+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.3 (2022-01-25)
+
+- Added option in volumes to prevent from showing up in upcoming releases
+- Fixed prices when creating a volume and not having set a default price in the series
+
 ## 1.2.2 (2022-01-22)
 
 - A local version of flasher is now used instead of loading it via a CDN